revert

+++ install-gui.sh
@@ -1,18 +1,20 @@
 #!/bin/bash
 set -e
+export NODE_OPTIONS="--max-old-space-size=3000"
+
 
 if [ -z "$VIRTUAL_ENV" ]; then
-  echo "This requires the flax python virtual environment."
+  echo "This requires the olive python virtual environment."
   echo "Execute '. ./activate' before running."
 	exit 1
 fi
 
 if [ "$(id -u)" = 0 ]; then
-  echo "The Flax Blockchain GUI can not be installed or run by the root user."
+  echo "The Olive Blockchain GUI can not be installed or run by the root user."
 	exit 1
 fi
 
-# Allows overriding the branch or commit to build in flax-blockchain-gui
+# Allows overriding the branch or commit to build in olive-blockchain-gui
 SUBMODULE_BRANCH=$1
 
 UBUNTU=false
@@ -23,20 +25,20 @@ if [ "$(uname)" = "Linux" ]; then
 		# Debian/Ubuntu
 		UBUNTU=true
 		sudo apt-get install -y npm nodejs libxss1
-	elif type yum &&  [ ! -f "/etc/redhat-release" ] && [ ! -f "/etc/centos-release" ] && [ ! -f /etc/rocky-release ]; then
+	elif type yum &&  [ ! -f "/etc/redhat-release" ] && [ ! -f "/etc/centos-release" ] && [ ! -f /etc/rocky-release ] && [ ! -f /etc/fedora-release ]; then
 		# AMZN 2
 		echo "Installing on Amazon Linux 2."
 		curl -sL https://rpm.nodesource.com/setup_12.x | sudo bash -
 		sudo yum install -y nodejs
-	elif type yum && [ ! -f /etc/rocky-release ] && [ -f /etc/redhat-release ] || [ -f /etc/centos-release ]; then
+	elif type yum && [ ! -f /etc/rocky-release ] && [ ! -f /etc/fedora-release ] && [ -f /etc/redhat-release ] || [ -f /etc/centos-release ]; then
 		# CentOS or Redhat
 		echo "Installing on CentOS/Redhat."
 		curl -sL https://rpm.nodesource.com/setup_12.x | sudo bash -
 		sudo yum install -y nodejs
-	elif type yum && [ -f /etc/rocky-release ]; then
+	elif type yum && [ -f /etc/rocky-release ] || [ -f /etc/fedora-release ]; then
                 # RockyLinux
-                echo "Installing on RockyLinux"
-                dnf module enable nodejs:12
+                echo "Installing on RockyLinux/Fedora"
+                sudo dnf module enable nodejs:12
                 sudo dnf install -y nodejs
         fi
 
@@ -78,7 +80,7 @@ if [ ! "$CI" ]; then
 	echo "Running git submodule update."
 	echo ""
 	git submodule update
-	cd flax-blockchain-gui
+	cd olive-blockchain-gui
 
 	if [ "$SUBMODULE_BRANCH" ];
 	then
@@ -98,6 +100,6 @@ else
 fi
 
 echo ""
-echo "Flax blockchain install-gui.sh completed."
+echo "Olive blockchain install-gui.sh completed."
 echo ""
-echo "Type 'cd flax-blockchain-gui' and then 'npm run electron &' to start the GUI."
+echo "Type 'cd olive-blockchain-gui' and then 'npm run electron &' to start the GUI."
